More authorization, fixed almost all authentication, some cascade delete things

+++ server/routers/questionsRouter.js
@@ -3,18 +3,28 @@ const router = Router();
 import db from "../database/createConnection.js";
 
 //################# GET questions ####################
-router.get("/api/questions", async (req, res) => {
-    const questions = await db.all("SELECT * FROM questions;");
- 
+router.get("/api/questions/surveys/:id", async (req, res) => {
+    const surveyid = Number(req.params.id);
+    const questions = await db.all(`SELECT * FROM questions WHERE surveyid = ?;` [surveyid]);
     res.send({ data: questions });
 });
 
 
 //################# POST question ####################
 router.post("/api/questions", async (req, res) => {
-    const { question, surveyid } = req.body;
-    const { changes } = await db.run(`INSERT INTO questions (question, surveyid) VALUES (?, ?);`, [question, surveyid]);
-    return res.send({ rowsAffected: changes });
+    if (req.session.loggedIn) {
+        const userid = req.session.userID;
+        const { question, selectedSurveyId } = req.body;
+        console.log(question, selectedSurveyId);
+        const surveyid = await db.get(`SELECT id FROM surveys WHERE id = ? and userid = ?`, [selectedSurveyId, userid]);
+        console.log(surveyid);
+        if (surveyid) {
+            const { changes } = await db.run(`INSERT INTO questions (question, surveyid) VALUES (?, ?);`, [question, surveyid.id]);
+            return res.send({ rowsAffected: changes });  
+        }
+        return res.send({ error: "wrong surveyid"});
+    }
+    res.send({error: "not logged in"});
 });
 
 
@@ -24,13 +34,22 @@ router.post("/api/questions", async (req, res) => {
 
 
 //################# DELETE question ####################
-router.delete("/api/questions/:id", async (req, res) => {
-    const questionid = Number(req.params.id);
-    const { changes } = await db.run(`DELETE FROM questions WHERE id = ?`, questionid);
-    if (changes !== 0) {
-        return res.send({ rowsDeleted: changes })
+router.delete("/api/questions/:questionid/surveys/:surveyid", async (req, res) => {
+    if (req.session.loggedIn) {
+        const userid = req.session.userID;
+        const selectedSurveyId = Number(req.params.surveyid);
+        const questionid = Number(req.params.questionid);
+        const surveyid = await db.get(`SELECT id FROM surveys WHERE id = ? and userid = ?`, [selectedSurveyId, userid]);
+        if (surveyid) {
+            const { changes } = await db.run(`DELETE FROM questions WHERE id = ?`, questionid);
+            if (changes !== 0) {
+                return res.send({ rowsDeleted: changes })
+            }
+            return res.send({error: `No product by id: ${questionid}`})
+        }
+        return res.send({ error: "wrong surveyid"});
     }
-    return res.send({error: `No product by id: ${questionid}`})
+    res.send({error: "not logged in"});
 });
 
 export default router;
